Update comments and README

+++ struct/ordered_set.cpp
@@ -6,8 +6,10 @@ using namespace __gnu_pbds;
 #define ll long long
 #define pii pair<int,int>
 #define pll pair<long long,long long>
-typedef tree<ll, null_type, less<ll>, rb_tree_tag, tree_order_statistics_node_update> ordered_set;
-typedef tree<ll, null_type, less_equal<ll>, rb_tree_tag, tree_order_statistics_node_update> ordered_multiset;
+typedef tree<int, null_type, less<int>, rb_tree_tag, tree_order_statistics_node_update> ordered_set;
+typedef tree<int, null_type, less_equal<int>, rb_tree_tag, tree_order_statistics_node_update> ordered_multiset;
+// find_by_order(k) returns to an iterator to the kth element
+// order_of_key(k) returns to the number of items strictly smaller than k
 // lower_bound and upper_bound should be swapped for indexed_multiset
 // currently cannot run on mac, need to use custom invocation
